Crate crab_runtime

The Crab runtime. This can be compiled with ``#[no_std]`, ready for Wasm.

Modules

api

constants
Constant values used within the runtime. A set of constant values used in Crab runtime.

Structs

BasicDeposit

BondingDurationInBlockNumber

BondingDurationInEra

Burn

CandidacyBond

CandidateDeposit

Cap

ChallengePeriod

ClaimsList

ConfigDepositBase

CouncilMotionDuration

DealWithFees

DesiredMembers

DesiredRunnersUp

DisabledValidatorsThreshold

EpochDuration

EthNetwork

EtherScanAPIKey

ExistentialDeposit

ExpectedBlockTime

FetchInterval

FieldDeposit

FriendDepositFactor

GenesisConfig

IndexDeposit

KtonProposalBondMinimum

MaxAdditionalFields

MaxFriends

MaxLockDuration

MaxNominatorRewardedPerValidator

MaxSignatories

MaxStrikes

MaxSubAccounts

MinVestedTransfer

MinimumPeriod

ModuleToIndex
Provides an implementation of ModuleToIndex to map a module to its index in the runtime.

MultisigDepositBase

MultisigDepositFactor

PeriodSpend

Prefix

ProposalBond

RecoveryDeposit

ReportLatency

RingProposalBondMinimum

RotationPeriod

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

SessionDuration

SessionKeys

SessionsPerEra

SlashDeferDuration

SpendPeriod

SubAccountDeposit

SubKeyPrefix

TargetBlockFullness

TechnicalMotionDuration

TermDuration
Daily council elections.

TipCountdown

TipFindersFee

TipReportDepositBase

TipReportDepositPerByte

TotalPower

TransactionBaseFee

TransactionByteFee

UncleGenerations

Version

VotingBond

WeightFeeCoefficient

WindowSize

WrongSideDeduction

Enums

Call

Event

Forcing
Mode of era-forcing.

Origin

StakerStatus
Indicates the initial status of the staker.

Constants

VERSION
Runtime version (Crab).

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version
Native version.

Type Definitions

Address
The address format for describing accounts.

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

CheckedExtrinsic
Extrinsic type that has already been checked.

Claims

ClaimsConfig

Council

CouncilConfig

ElectionsPhragmen

EthBacking

EthBackingConfig

EthOffchain

EthRelay

EthRelayConfig

Executive
Executive: handles dispatch to the various modules.

FinalityTracker

Grandpa

GrandpaConfig

Header
Block header type as expected by this runtime.

HeaderMMR

Historical

Identity

ImOnline

ImOnlineConfig

Indices

IndicesConfig

Kton

KtonConfig

Offences

RandomnessCollectiveFlip

Recovery

Session

SessionConfig

SignedBlock
A Block signed with a Justification

SignedExtra
The SignedExtension to the basic transaction logic.

SignedPayload
The payload being signed in transactions.

Society

Staking

StakingConfig

Sudo

SudoConfig

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalMembership

TechnicalMembershipConfig

Timestamp

TransactionPayment

Treasury

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.

Utility

Vesting

VestingConfig